agent/requirements.txt
agno==1.6.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
arxiv==2.2.0
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Crawl4AI==0.6.3
cryptography==45.0.4
cssselect==1.3.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==8.0.4
fake-http-header==0.3.5
fake-useragent==2.2.0
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.21.1
google-geo-type==0.3.13
google-maps-places==0.2.1
googleapis-common-protos==1.70.0
googlemaps==4.10.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.9
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.3.0
ollama==0.5.1
openai==1.90.0
packaging==25.0
pandas==2.3.0
pillow==10.4.0
playwright==1.52.0
portalocker==2.10.1
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pymongo==4.13.2
pyOpenSSL==25.1.0
pypdf==5.6.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
tenacity==8.5.0
tf-playwright-stealth==1.2.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
